# copr: optimize limit operator next_batch method