--- a/components/tidb_query_executors/src/limit_executor.rs
+++ b/components/tidb_query_executors/src/limit_executor.rs
@@ -34,10 +34,11 @@ impl<Src: BatchExecutor> BatchExecutor for BatchLimitExecutor<Src> {
 
     #[inline]
     fn next_batch(&mut self, scan_rows: usize) -> BatchExecuteResult {
-        let mut real_scan_rows = scan_rows;
-        if self.is_src_scan_executor {
-            real_scan_rows = std::cmp::min(scan_rows, self.remaining_rows);
-        }
+        let real_scan_rows = if self.is_src_scan_executor {
+            std::cmp::min(scan_rows, self.remaining_rows)
+        } else {
+            scan_rows
+        };
         let mut result = self.src.next_batch(real_scan_rows);
         if result.logical_rows.len() < self.remaining_rows {
             self.remaining_rows -= result.logical_rows.len();
@@ -79,6 +80,7 @@ mod tests {
     use tidb_query_datatype::FieldTypeTp;
 
     use crate::util::mock_executor::MockExecutor;
+    use crate::util::mock_executor::MockScanExecutor;
     use tidb_query_datatype::codec::batch::LazyBatchColumnVec;
     use tidb_query_datatype::codec::data_type::VectorValue;
     use tidb_query_datatype::expr::EvalWarnings;
@@ -247,5 +249,29 @@ mod tests {
         assert_eq!(&r.logical_rows, &[0, 4]);
         assert_eq!(r.physical_columns.rows_len(), 5);
         assert!(r.is_drained.unwrap());
+    }
+
+    #[test]
+    fn test_src_exec_is_scan() {
+        let schema = vec![FieldTypeTp::LongLong.into()];
+        let rows = (0..1024).collect();
+        let src_exec = MockScanExecutor::new(rows, schema);
+
+        let mut exec = BatchLimitExecutor::new(src_exec, 5, true).unwrap();
+        let r = exec.next_batch(100);
+        assert_eq!(r.logical_rows, &[0, 1, 2, 3, 4]);
+        let r = exec.next_batch(2);
+        assert_eq!(r.is_drained.unwrap(), true);
+
+        let schema = vec![FieldTypeTp::LongLong.into()];
+        let rows = (0..1024).collect();
+        let src_exec = MockScanExecutor::new(rows, schema);
+        let mut exec = BatchLimitExecutor::new(src_exec, 1024, true).unwrap();
+        for _i in 0..1023 {
+            let r = exec.next_batch(1);
+            assert_eq!(r.is_drained.unwrap(), false);
+        }
+        let r = exec.next_batch(1);
+        assert_eq!(r.is_drained.unwrap(), true);
     }
 }