--- a/components/tidb_query_executors/src/runner.rs
+++ b/components/tidb_query_executors/src/runner.rs
@@ -159,6 +159,7 @@ pub fn build_executors<S: Storage + 'static>(
 
     let mut executor: Box<dyn BatchExecutor<StorageStats = S::Statistics>>;
     let mut summary_slot_index = 0;
+    let mut is_scan_executor = true;
 
     match first_ed.get_tp() {
         ExecType::TypeTableScan => {
@@ -286,8 +287,12 @@ pub fn build_executors<S: Storage + 'static>(
                 EXECUTOR_COUNT_METRICS.batch_limit.inc();
 
                 Box::new(
-                    BatchLimitExecutor::new(executor, ed.get_limit().get_limit() as usize)?
-                        .collect_summary(summary_slot_index),
+                    BatchLimitExecutor::new(
+                        executor,
+                        ed.get_limit().get_limit() as usize,
+                        is_scan_executor,
+                    )?
+                    .collect_summary(summary_slot_index),
                 )
             }
             ExecType::TypeTopN => {
@@ -321,6 +326,7 @@ pub fn build_executors<S: Storage + 'static>(
             }
         };
         executor = new_executor;
+        is_scan_executor = false;
     }
 
     Ok(executor)