--- a/components/tidb_query_executors/src/runner.rs
+++ b/components/tidb_query_executors/src/runner.rs
@@ -159,7 +159,9 @@ pub fn build_executors<S: Storage + 'static>(
 
     let mut executor: Box<dyn BatchExecutor<StorageStats = S::Statistics>>;
     let mut summary_slot_index = 0;
-    let mut is_scan_executor = true;
+    // Limit executor use this flag to check if its src is table/index scan.
+    // Performance enhancement for plan like: limit 1 -> table/index scan.
+    let mut is_src_scan_executor = true;
 
     match first_ed.get_tp() {
         ExecType::TypeTableScan => {
@@ -290,7 +292,7 @@ pub fn build_executors<S: Storage + 'static>(
                     BatchLimitExecutor::new(
                         executor,
                         ed.get_limit().get_limit() as usize,
-                        is_scan_executor,
+                        is_src_scan_executor,
                     )?
                     .collect_summary(summary_slot_index),
                 )
@@ -326,7 +328,7 @@ pub fn build_executors<S: Storage + 'static>(
             }
         };
         executor = new_executor;
-        is_scan_executor = false;
+        is_src_scan_executor = false;
     }
 
     Ok(executor)