Signed-off-by: guo-shaoge <shaoge1994@163.com>

<!--
Thank you for contributing to TiKV!

If you haven't already, please read TiKV's [CONTRIBUTING](https://github.com/tikv/tikv/blob/master/CONTRIBUTING.md) document.

If you're unsure about anything, just ask; somebody should be along to answer within a day or two.

PR Title Format:
1. module [, module2, module3]: what's changed
2. *: what's changed

If you want to open the **Challenge Program** pull request, please use the following template:
https://raw.githubusercontent.com/tikv/.github/master/.github/PULL_REQUEST_TEMPLATE/challenge-program.md
You can use it with query parameters: https://github.com/tikv/tikv/compare/master...${you branch}?template=challenge-program.md
-->

### What problem does this PR solve?

Issue Number: close #9828 <!-- REMOVE this line if no issue to close -->

Problem Summary: Performance enhancement for Limit operator. When chall child's next_batch(scan_rows), default scan_rows is 32, but if limit number is 1 or less, we scan more rows than we actually need.

### What is changed and how it works?
What's Changed: change LimitBatchExecutor::next_batch(). See files changed, it's a small change.

### Related changes


- Need to cherry-pick to the release branch

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- performance test
- unitest


Side effects

- performance enhancement for limit operator. [here](https://github.com/tikv/tikv/pull/9851#issuecomment-803539262)


### Release note <!-- bugfixes or new feature need a release note -->
- copr: performance enhancement for limit operator