@guo-shaoge 2021-03-21
环境：
    4核 AMD Ryzen 5 PRO 4650U with Radeon Graphics
    4G内存
    3个TiDB, 1个TiKV, 1个PD

测试SQL：
    use test;
    drop table if exists t;
    create table t (gjt_col bigint primary key, val varchar(1024));
    load data local infile '/home/guo-shaoge/work/limit_perf/row_1000000_col_2' into table t fields terminated by ' ';
    select * from t order by gjt_col limit 1;

测试方法：
    客户端并发度30，以round robin方式将不同线程的请求平均发送给3个TiDB。每个TiDB收到请求后，会将select语句循环2000次

    统计的时间就是2000次执行的时间，测试时TiKV的cpu被压满，CPU占用基本到400%。测试结论如下

结果：
    Limit算子修改前: 执行时间在49.9s到1min之间，平均时间为 1min

    修改后：30个线程的执行时间在27.9s到29.7s之间，平均时间为 28s

可以看到，只有Limit算子情况下，性能提升比较大，但是带有selection的SQL有回退

测试SQL，schema等信息同上一个测试一样:
    select gjt_col from t where val > 3211489453760402549 limit 1;

测试方法：
    跟上面的测试方法一样，但是并发度改为10

结果：
   Limit算子修改前: 执行时间在1min8s到1m40s之间，平均时间为 1min25s

    修改后：30个线程的执行时间在42s到1min之间，平均时间为 55s