@guo-shaoge 2021-03-21
But only change Limit operator will affect performance of  **selection** operator. Check the following test.

enviroments:

        4 cores. AMD Ryzen 5 PRO 4650U with Radeon Graphics
        4G memory
        3 TiDB, 1 TiKV, 1 PD

SQL(same schema and input data with above test):

    ('3211489453760402549' is a random choice, There are 733502 rows out of 1000000 satify this filter)
    SQL1: select gjt_col from t where val > '3211489453760402549' limit 1;

    ('999995990325667965' is the max value for column `val`, so table scan needs to scan all rows of table `t`)
    SQL2: select gjt_col from t where val > '999995990325667965' limit 1;

    ('100000152253498439' is the min value for column `val`, so table scan only need to scan small number of rows)
    SQL3: select gjt_col from t where val >= '100000152253498439' limit 1

how to test:

    SQL1: Same as above. But client will use 10 goroutines(Because workload is too heavy, cost too much time).

    SQL2: TiDB will execute each request 20 times, and client will only have 3 goroutines(workload is too heavy).

    SQL3: Save as above. 2000 times for each request, and client has 30 goroutines.

SQL1 result:
    
    before: exec duratoin for 30 goroutines: [42s, 1min]，avg: 55s

    after: 1min8s, 1m40s，avg: 1min25s (**regression here!**)

SQL2 result:

    before: avg: 7s

    after: avg: 18s (**regression here!**)

SQL3 result:

    before: avg: 55s

    after: avg: 53s

So need to also change implementation of selection's next_batch() method.