@guo-shaoge 2021-03-21
> Can you post a summary of the result (images are better) in the PR description? I (and many other contributors) don't have the permission to see the doc.

Ok. Check [limit operator enhancement](https://github.com/tikv/tikv/pull/9851#issuecomment-803539262)  and [select operation regression](https://github.com/tikv/tikv/pull/9851#issuecomment-803578571)